+++ spring-cloud-starter-tencent-polaris-config/src/main/java/com/tencent/cloud/polaris/config/PolarisConfigAutoConfiguration.java
@@ -18,11 +18,14 @@
 
 package com.tencent.cloud.polaris.config;
 
-import com.tencent.cloud.polaris.config.adapter.PolarisPropertySourceAutoRefresher;
+import com.tencent.cloud.polaris.config.adapter.AffectedConfigurationPropertiesRebinder;
+import com.tencent.cloud.polaris.config.adapter.PolarisConfigPropertyRefresher;
 import com.tencent.cloud.polaris.config.adapter.PolarisPropertySourceManager;
-import com.tencent.cloud.polaris.config.adapter.SmartConfigurationPropertiesRebinder;
+import com.tencent.cloud.polaris.config.adapter.PolarisRefreshAffectedContextRefresher;
+import com.tencent.cloud.polaris.config.adapter.PolarisRefreshEntireContextRefresher;
 import com.tencent.cloud.polaris.config.annotation.PolarisConfigAnnotationProcessor;
-import com.tencent.cloud.polaris.config.condition.ConditionalOnNonDefaultBehavior;
+import com.tencent.cloud.polaris.config.condition.ConditionalOnConfigReflectEnabled;
+import com.tencent.cloud.polaris.config.condition.ConditionalOnReflectRefreshType;
 import com.tencent.cloud.polaris.config.config.PolarisConfigProperties;
 import com.tencent.cloud.polaris.config.listener.PolarisConfigChangeEventListener;
 import com.tencent.cloud.polaris.config.spring.annotation.SpringValueProcessor;
@@ -30,6 +33,7 @@ import com.tencent.cloud.polaris.config.spring.property.PlaceholderHelper;
 import com.tencent.cloud.polaris.config.spring.property.SpringValueRegistry;
 import com.tencent.cloud.polaris.context.ConditionalOnPolarisEnabled;
 
+import org.springframework.boot.autoconfigure.AutoConfigureBefore;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.autoconfigure.condition.SearchStrategy;
@@ -50,17 +54,6 @@ import org.springframework.context.annotation.Configuration;
 public class PolarisConfigAutoConfiguration {
 
 	@Bean
-	public PolarisPropertySourceAutoRefresher polarisPropertySourceAutoRefresher(
-			PolarisConfigProperties polarisConfigProperties,
-			PolarisPropertySourceManager polarisPropertySourceManager,
-			SpringValueRegistry springValueRegistry,
-			PlaceholderHelper placeholderHelper,
-			ContextRefresher contextRefresher) {
-		return new PolarisPropertySourceAutoRefresher(polarisConfigProperties,
-				polarisPropertySourceManager, springValueRegistry, placeholderHelper, contextRefresher);
-	}
-
-	@Bean
 	public PolarisConfigAnnotationProcessor polarisConfigAnnotationProcessor() {
 		return new PolarisConfigAnnotationProcessor();
 	}
@@ -71,28 +64,46 @@ public class PolarisConfigAutoConfiguration {
 	}
 
 	@Bean
-	public SpringValueRegistry springValueRegistry() {
-		return new SpringValueRegistry();
-	}
-
-	@Bean
-	public PlaceholderHelper placeholderHelper() {
-		return new PlaceholderHelper();
-	}
-
-	@Bean
-	public SpringValueProcessor springValueProcessor(PlaceholderHelper placeholderHelper, SpringValueRegistry springValueRegistry, PolarisConfigProperties polarisConfigProperties) {
-		return new SpringValueProcessor(placeholderHelper, springValueRegistry, polarisConfigProperties);
+	@ConditionalOnMissingBean(search = SearchStrategy.CURRENT)
+	@ConditionalOnReflectRefreshType
+	public ConfigurationPropertiesRebinder affectedConfigurationPropertiesRebinder(
+			ConfigurationPropertiesBeans beans) {
+		return new AffectedConfigurationPropertiesRebinder(beans);
 	}
 
 	@Bean
 	@ConditionalOnMissingBean(search = SearchStrategy.CURRENT)
-	@ConditionalOnNonDefaultBehavior
-	public ConfigurationPropertiesRebinder smartConfigurationPropertiesRebinder(
-			ConfigurationPropertiesBeans beans) {
-		// If using default behavior, not use SmartConfigurationPropertiesRebinder.
-		// Minimize te possibility of making mistakes.
-		return new SmartConfigurationPropertiesRebinder(beans);
+	public PolarisConfigPropertyRefresher polarisRefreshContextPropertySourceAutoRefresher(PolarisConfigProperties polarisConfigProperties,
+			PolarisPropertySourceManager polarisPropertySourceManager, ContextRefresher contextRefresher) {
+		return new PolarisRefreshEntireContextRefresher(polarisConfigProperties, polarisPropertySourceManager, contextRefresher);
 	}
 
+	@ConditionalOnConfigReflectEnabled
+	@Configuration(proxyBeanMethods = false)
+	@AutoConfigureBefore(PolarisConfigAutoConfiguration.class)
+	public static class PolarisReflectRefresherAutoConfiguration {
+		@Bean
+		public SpringValueRegistry springValueRegistry() {
+			return new SpringValueRegistry();
+		}
+
+		@Bean
+		public PlaceholderHelper placeholderHelper() {
+			return new PlaceholderHelper();
+		}
+
+		@Bean
+		public SpringValueProcessor springValueProcessor(PlaceholderHelper placeholderHelper,
+				SpringValueRegistry springValueRegistry, PolarisConfigProperties polarisConfigProperties) {
+			return new SpringValueProcessor(placeholderHelper, springValueRegistry, polarisConfigProperties);
+		}
+
+		@Bean
+		public PolarisConfigPropertyRefresher polarisReflectPropertySourceAutoRefresher(PolarisConfigProperties polarisConfigProperties,
+				PolarisPropertySourceManager polarisPropertySourceManager, SpringValueRegistry springValueRegistry,
+				PlaceholderHelper placeholderHelper) {
+			return new PolarisRefreshAffectedContextRefresher(polarisConfigProperties, polarisPropertySourceManager,
+					springValueRegistry, placeholderHelper);
+		}
+	}
 }