+++ CHANGELOG.md
@@ -25,3 +25,4 @@
 - [Feature：support pushGateway push metrics](https://github.com/Tencent/spring-cloud-tencent/pull/529)
 - [Feature: remove location metadata](https://github.com/Tencent/spring-cloud-tencent/pull/536)
 - [add feature-env plugin & add spring cloud gateway staining plugin](https://github.com/Tencent/spring-cloud-tencent/pull/533)
+- [feature:add @ConditionalOnConfigReflectEnabled annotation](https://github.com/Tencent/spring-cloud-tencent/pull/539)